Adding travis build
