+++ .travis.yml
@@ -1,0 +1,10 @@
+language: python
+python:
+  - "2.7"
+  - "3.3"
+  - "3.4"
+cache: pip
+install:
+  - pip install -r requirements.txt
+script:
+  - py.test
