+++ requirements.txt
@@ -1,0 +1,2 @@
+pytest
+tox-travis
