+++ sorts/mergesort.py
@@ -1,5 +1,3 @@
-import unittest
-
 def mergesort(array):
     if len(array) < 2:
         return array
@@ -20,14 +18,3 @@ def merge(left, right):
             j += 1
     combined += left[i:] or right[j:]
     return combined
-
-class TestMergesort(unittest.TestCase):
-    def test(self):
-        self.assertEqual(mergesort([5, 5, 6, 123, 5464, 2, 0, -1]), [-1, 0, 2, 5, 5, 6, 123, 5464])
-        self.assertEqual(mergesort([5, 6]), [5, 6])
-        self.assertEqual(mergesort([5]), [5])
-        self.assertEqual(mergesort([6, 5]), [5, 6])
-        self.assertEqual(mergesort([]), [])
-        self.assertEqual(mergesort([5, 5, 5, 5, 5]), [5, 5, 5, 5, 5])
-
-unittest.main()
