+++ sorts/quicksort.py
@@ -1,5 +1,3 @@
-import unittest
-
 def quicksort(array):
     inplace_quicksort(array, 0, len(array) - 1)
     return array
@@ -22,14 +20,3 @@ def inplace_quicksort(array, start, end):
     array[lptr], array[end] = array[end], array[lptr]
     inplace_quicksort(array, start, lptr - 1)
     inplace_quicksort(array, lptr + 1, end)
-
-class TestQuicksort(unittest.TestCase):
-    def test(self):
-        self.assertEqual(quicksort([5, 6, 123, 5464, 2, 0, -1]), [-1, 0, 2, 5, 6, 123, 5464])
-        self.assertEqual(quicksort([5, 6]), [5, 6])
-        self.assertEqual(quicksort([5]), [5])
-        self.assertEqual(quicksort([6, 5]), [5, 6])
-        self.assertEqual(quicksort([]), [])
-        self.assertEqual(quicksort([5, 5, 5, 5, 5]), [5, 5, 5, 5, 5])
-
-unittest.main()
